refactor: download delegesini DioMixin.clone() ile kur
Delege elle `Dio()` kurularak oluşturuluyordu. Bunun iki eksiği vardı:
`Dio()` kurucusu varsayılan ImplyContentTypeInterceptor'ı ekliyor (manager
onu taşımıyor) ve transformer kopyalanmıyordu.

`DioMixin.clone()` base options, interceptor zinciri, transformer ve
httpClientAdapter'ı birebir taşır, fazladan interceptor'ı da kaldırır.
Yaklaşım PR #135'ten alındı.

Faz 1 testleri de konularına ayrıldı: phase1_cleanup_test.dart yerine
dio_delegation_test.dart ve network_check_platform_test.dart.


# Conflicts:
#	packages/vexana/test/unit/phase1_cleanup_test.dart
